requirements.txt
# Dependências para otimização de imagens
Pillow>=10.0.0
tqdm>=4.66.0
